server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
scikit-learn==1.3.2
llama-cpp-python>=0.2.0
huggingface-hub>=0.16.4
